requirements_vercel.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
google-generativeai==0.8.5
gunicorn==23.0.0
line-bot-sdk==3.13.0
openai==1.58.1
psycopg2-binary==2.9.10
requests==2.32.3
werkzeug==3.0.6
email-validator==2.2.0